Seperated out all of the other services
